requirements.txt
alembic==1.13.3
babel==2.16.0
docxcompose==1.4.0
docxtpl==0.18.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
pillow==10.4.0
protobuf==5.28.2
psycopg2==2.9.10
python-docx==1.1.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
